Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/parsing/border-image-shorthand.sub-expected.txt


FAIL e.style['border-image'] = "none" should set border-image-outset assert_equals: border-image-outset should be canonical expected "0" but got "initial"
FAIL e.style['border-image'] = "none" should set border-image-repeat assert_equals: border-image-repeat should be canonical expected "stretch" but got "initial"
FAIL e.style['border-image'] = "none" should set border-image-slice assert_equals: border-image-slice should be canonical expected "100%" but got "initial"
PASS e.style['border-image'] = "none" should set border-image-source
FAIL e.style['border-image'] = "none" should set border-image-width assert_equals: border-image-width should be canonical expected "1" but got "initial"
PASS e.style['border-image'] = "none" should not set unrelated longhands
FAIL e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should set border-image-outset assert_equals: border-image-outset should be canonical expected "0" but got "initial"
FAIL e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should set border-image-repeat assert_equals: border-image-repeat should be canonical expected "stretch" but got "initial"
PASS e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should set border-image-slice
PASS e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should set border-image-source
FAIL e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should set border-image-width assert_equals: border-image-width should be canonical expected "1" but got "initial"
PASS e.style['border-image'] = "url(\"http://localhost/\") 1 2 3 4 fill" should not set unrelated longhands
FAIL e.style['border-image'] = "repeat round" should set border-image-outset assert_equals: border-image-outset should be canonical expected "0" but got "initial"
PASS e.style['border-image'] = "repeat round" should set border-image-repeat
FAIL e.style['border-image'] = "repeat round" should set border-image-slice assert_equals: border-image-slice should be canonical expected "100%" but got "initial"
FAIL e.style['border-image'] = "repeat round" should set border-image-source assert_equals: border-image-source should be canonical expected "none" but got "initial"
FAIL e.style['border-image'] = "repeat round" should set border-image-width assert_equals: border-image-width should be canonical expected "1" but got "initial"
PASS e.style['border-image'] = "repeat round" should not set unrelated longhands
FAIL e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should set border-image-outset assert_equals: border-image-outset should be canonical expected "0" but got "initial"
FAIL e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should set border-image-repeat assert_equals: border-image-repeat should be canonical expected "stretch" but got "initial"
PASS e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should set border-image-slice
PASS e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should set border-image-source
FAIL e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should set border-image-width assert_equals: border-image-width should be canonical expected "1" but got "initial"
PASS e.style['border-image'] = "url(\"http://localhost/\") fill 1 2% 3 4%" should not set unrelated longhands
PASS e.style['border-image'] = "1 2% 3 4% / 5% / 6" should set border-image-outset
FAIL e.style['border-image'] = "1 2% 3 4% / 5% / 6" should set border-image-repeat assert_equals: border-image-repeat should be canonical expected "stretch" but got "initial"
PASS e.style['border-image'] = "1 2% 3 4% / 5% / 6" should set border-image-slice
FAIL e.style['border-image'] = "1 2% 3 4% / 5% / 6" should set border-image-source assert_equals: border-image-source should be canonical expected "none" but got "initial"
PASS e.style['border-image'] = "1 2% 3 4% / 5% / 6" should set border-image-width
PASS e.style['border-image'] = "1 2% 3 4% / 5% / 6" should not set unrelated longhands
